[FIX] PdfMerger -> PdfProcessor
